+++ pyluna-pathology/luna/pathology/cli/generate_tiles.py
@@ -9,9 +9,6 @@ logger = logging.getLogger('generate_tiles')
 
 from luna.common.utils import cli_runner
 
-import pandas as pd
-from tqdm import tqdm
-
 _params_ = [('input_slide_image', str), ('output_dir', str), ('tile_size', int), ('batch_size', int), ('requested_magnification', float), ('num_cores', int)]
 
 @click.command()
@@ -29,37 +26,37 @@ _params_ = [('input_slide_image', str), ('output_dir', str), ('tile_size', int),
 @click.option('-m', '--method_param_path', required=False,
               help='path to a metadata json/yaml file with method parameters to reproduce results')
 def cli(**cli_kwargs):
-    """ Run a model with a specific pre-transform for all tiles in a slide (tile_images)
+    """Rasterize a slide into smaller tiles
+    
+    Tiles are saved in the whole-slide tiles binary format (tiles.pil), and the corresponding manifest/header file (tiles.csv) is also generated
+
+    Neccessary data for the manifest file are: 
+    address, x_coord, y_coord, full_resolution_tile_size, tile_image_binary, tile_image_length, tile_image_size_xy, and tile_image_size_mode
 
     \b
     Inputs:
-        input_slide_image: path to slide tiles (.svs)
+        input_slide_image: slide image (.svs)
     \b
     Outputs:
         slide_tiles
     \b
     Example:
-        infer_tiles tiles/slide-100012/tiles
+        generate_tiles 10001.svs
+            -nc 8 -rts 244 -rmg 10 -bx 200
+            -o 10001/tiles
     """
     cli_runner( cli_kwargs, _params_, generate_tiles)
 
+import pandas as pd
+from tqdm import tqdm
 import openslide
-from luna.pathology.common.preprocess import get_scale_factor_at_magnfication, get_full_resolution_generator, coord_to_address, address_to_coord
 import itertools
 from pathlib import Path
 
 from concurrent.futures import ProcessPoolExecutor, as_completed
 
-def grouper(iterable, n, fillvalue=None):
-    args = [iter(iterable)] * n
-    return [[entry for entry in iterable if entry is not None]
-            for iterable in itertools.zip_longest(*args, fillvalue=fillvalue)]
-
-def get_tile_bytes(indices, input_slide_image, full_resolution_tile_size, tile_size ):
-    slide = openslide.OpenSlide(str(input_slide_image))
-    full_generator, full_level = get_full_resolution_generator(slide, tile_size=full_resolution_tile_size)
-    return [(index, full_generator.get_tile(full_level, address_to_coord(index)).resize((tile_size,tile_size)).tobytes()) for index in indices]
-
+from luna.pathology.common.utils import get_tile_bytes, get_scale_factor_at_magnfication, get_full_resolution_generator, coord_to_address
+from luna.common.utils import grouper
 def generate_tiles(input_slide_image, tile_size, requested_magnification, output_dir, num_cores, batch_size):
     """Rasterize a slide into smaller tiles
     
@@ -100,13 +97,15 @@ def generate_tiles(input_slide_image, tile_size, requested_magnification, output
     logger.info("tiles x %s, tiles y %s", tile_x_count, tile_y_count)
 
     # populate address, coordinates
-    address_raster = [{"address": coord_to_address(address, requested_magnification), "x_coord": (address[0]) * full_resolution_tile_size, "y_coord": (address[1]) * full_resolution_tile_size}
+    address_raster = [{
+        "address": coord_to_address(address, requested_magnification), 
+        "x_coord": (address[0]) * full_resolution_tile_size, 
+        "y_coord": (address[1]) * full_resolution_tile_size}
                       for address in itertools.product(range(1, tile_x_count-1), range(1, tile_y_count-1))]
+                      
     logger.info("Number of tiles in raster: %s", len(address_raster))
 
     df = pd.DataFrame(address_raster).set_index("address")
-
-    df.loc[:, 'tile_image_offset']  = None
 
     output_binary_file = f"{output_dir}/{slide_name}.tiles.pil"
     output_header_file = f"{output_dir}/{slide_name}.tiles.csv"
@@ -115,17 +114,20 @@ def generate_tiles(input_slide_image, tile_size, requested_magnification, output
     offset = 0
     counter = 0
     logger.info(f"Now generating tiles with num_cores={num_cores} and batch_size={batch_size}!")
+
+    address_offset = []
     with ProcessPoolExecutor(num_cores) as executor:
         out = [executor.submit(get_tile_bytes, index, input_slide_image, full_resolution_tile_size, tile_size ) for index in grouper(df.index, batch_size)]
         for future in tqdm(as_completed(out), file=sys.stdout, total=len(out)):
             for index, tile in future.result(): 
                 fp.write( tile )
                 
-                df.loc[index, "tile_image_offset"]   = int(offset)
-
+                address_offset.append ((index, int(offset)))
                 offset += len(tile)
                 counter+=1
     fp.close()
+
+    df = df.join(pd.DataFrame(address_offset, columns=['address', 'tile_image_offset']).set_index('address'))
 
     df.loc[:, 'full_resolution_tile_size'] = full_resolution_tile_size
     df.loc[:, 'tile_image_binary']  = output_binary_file
