+++ pyluna-pathology/luna/pathology/cli/visualize_tile_labels_png.py
@@ -9,9 +9,6 @@ logger = logging.getLogger('visualize_tiles_png')
 
 from luna.common.utils import cli_runner
 
-import pandas as pd
-from tqdm import tqdm
-
 from typing import List
 
 _params_ = [('input_slide_image', str), ('input_slide_tiles', str), ('plot_labels', List[str]), ('requested_magnification', float), ('output_dir', str)]
@@ -49,14 +46,11 @@ def cli(**cli_kwargs):
 
 import openslide
 
-from luna.pathology.common.preprocess import get_downscaled_thumbnail, get_scale_factor_at_magnfication
-from luna.pathology.common.preprocess import get_tile_color
-from skimage.draw import rectangle_perimeter
+from luna.pathology.common.utils import visualize_tiling_scores, get_downscaled_thumbnail, get_scale_factor_at_magnfication
+import pandas as pd
 
 from pathlib import Path
 from PIL import Image
-import numpy as np
-
 def visualize_tiles(input_slide_image, input_slide_tiles, requested_magnification, plot_labels, output_dir):
     """Generate nice tile markup images with continuous or discrete tile scores
 
@@ -96,41 +90,6 @@ def visualize_tiles(input_slide_image, input_slide_tiles, requested_magnificatio
     properties = {'data': output_dir}
 
     return properties
-
-
-def visualize_tiling_scores(df:pd.DataFrame, thumbnail_img:np.ndarray, scale_factor:float,
-        score_type_to_visualize:str) -> np.ndarray:
-    """visualize tile scores
-    
-    draws colored boxes around tiles to indicate the value of the score 
-
-    Args:
-        df (pd.DataFrame): input dataframe
-        thumbnail_img (np.ndarray): input tile 
-        tile_size (int): tile width/length
-        score_type_to_visualize (str): column name from data frame
-    
-    Returns:
-        np.ndarray: new thumbnail image with boxes around tiles passing indicating the
-        value of the score
-    """
-
-    assert isinstance(thumbnail_img, np.ndarray)
-
-    for _, row in tqdm(df.iterrows()):
-
-        if 'regional_label' in row and pd.isna(row.regional_label): continue
-
-        start = (row.y_coord / scale_factor, row.x_coord / scale_factor)  # flip because OpenSlide uses (column, row), but skimage, uses (row, column)
-
-        rr, cc = rectangle_perimeter(start=start, extent=(row.full_resolution_tile_size/ scale_factor, row.full_resolution_tile_size/ scale_factor), shape=thumbnail_img.shape)
-        
-        # set color based on intensity of value instead of black border (1)
-        score = row[score_type_to_visualize]
-
-        thumbnail_img[rr, cc] = get_tile_color(score)
-    
-    return thumbnail_img
 
 if __name__ == "__main__":
     cli()
